Complete implementation with change to store head node in array
Store head node of linked list not the linked list itself I don't need
any of the data from it. The head node is enough to access the rest of
the nodes. Head node through property accesses the rest.
Otherwise this caused incosistency in accessing the head.value vs value
for a plain node.